% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{api_roclet}
\alias{api_roclet}
\title{Roclet: make API file.}
\usage{
api_roclet()
}
\description{
This roclet automates the production of an \code{API} file that describes
the exported interface of a package.
}
