% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{extract_api}
\alias{extract_api}
\title{Get package data}
\usage{
extract_api(path = ".", targets = character())
}
\arguments{
\item{path}{package root}

\item{targets}{character vector, function call targets to find}
}
\value{
Named list with components:
\itemize{
\item \code{name} is the package name.
\item \code{version} is the vesion of the package.
\item \code{targets} is a character vector, the names of the environments where
each target was found.
\item \code{functions} is a list of functions, with source references.
\item \code{exports} is a character vector of exported objects.
\item \code{s3_methods} is a character vector of declared S3 methods.
\item \code{imports} is a named list of environment names, one for each import.
}
}
\description{
Parse and evaluate code from a package, and store it in an
environment, so that we can access it later without parsing it again.
}
\details{
We need to evaluate code to get the functions, which might require
loading some packages, so we perform all this in a separate R process,
using the \code{callr} package.

While we are at it, we also check the imports of the package,
and return where each function was imported from.
}
\keyword{internal}
