% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-defs.R
\name{function_defs}
\alias{function_defs}
\title{Table of function definitions}
\usage{
function_defs(funcs, exports)
}
\arguments{
\item{path}{path to package root}
}
\value{
Data frame with columns: \code{name}, \code{file}, \code{line}, \code{col}.
}
\description{
Table of function definitions
}
\keyword{internal}
