% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_package}
\alias{map_package}
\title{Create the function call map of an R package}
\usage{
map_package(path = ".")
}
\arguments{
\item{path}{path to an R package root}
}
\value{
List with elements:
\itemize{
\item \code{name}: package name,
\item \code{exports}: character vector of exported funtions and operators.
\item \code{defs}: data frame of function definitions.
\item \code{calls}: data frame of function calls.
}

Columns in \code{defs}:
\itemize{
\item \code{name}: function name,
\item \code{file}: file path where it was defined, relative to the package root.
\item \code{line1}: line of start position of function definition.
\item \code{col1}: column of start position.
\item \code{line2}: line of end position.
\item \code{col2}: column of end position.
\item \code{exported}: whether the function is exported or not.
}

Columns in \code{calls}:
\itemize{
\item \code{file}: file path of the call, relative to the package root.
\item \code{from}: name of calling function, or empty string if called from
outside of a function.
\item \code{to}: called function, in the \code{package::function} form.
\item \code{type}: call type, regular functions calls are \code{call}.
\item \code{line1}: line of the beginning of the call.
\item \code{col1}: column of the beginning of the call.
\item \code{line1}: line of the end of the call.
\item \code{col2}: column of the beginning of the call.
\item \code{str}: the actual code, text of the function call, usually just the
called function's name.
}
}
\description{
Find all functions and function calls in a package. This requires the
loading of the package (in another R session, so the current R session
is unaffected by this), which means that all depended and imported
packages must be available as well.
}
