% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ functionmap2 }
\alias{ functionmap2 }
\alias{ functionmap2 }
\title{ Map Function Calls }
\description{
Create the map function calls in a package, including calls to imported
  packages.
}
