% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_expressions}
\alias{parse_expressions}
\title{Parse all files of a package}
\usage{
parse_expressions(path)
}
\arguments{
\item{path}{package root}
}
\value{
List of expressions, one for each file, names are full
file names.
}
\description{
This includes all (unevaluated) code of the package. We use it
to find all function calls.
}
\keyword{internal}

