% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_package}
\alias{prepare_package}
\title{Get package data}
\usage{
prepare_package(path, targets)
}
\arguments{
\item{path}{package root}

\item{targets}{character vector, function call targets to find}
}
\value{
List with components \code{name}, \code{version}, \code{targets}, \code{functions},
\code{exports}, \code{imports}.
\code{name} is the package name.
\code{version} is the vesion of the package.
\code{targets} is a character vector, the names of the environments where
each target was found.
\code{functions} is a list of functions, with source references.
\code{exports} is a character vector of exported objects.
\code{imports} is a named list of environment names, one for each import.
}
\description{
Parse and evaluate code from a package, and store it in an
environment, so that we can access it later without parsing it again.
}
\details{
We need to evaluate code to get the functions, which might require
loading some packages, so we perform all this in a separate R process,
using the \code{callr} package.

While we are at it, we also check the imports of the package,
and return where each function was imported from.
}
\keyword{internal}

